Feat: Add ability to merge standard user bulk pull requests
Creates SHA sum of GitHub registered username and email address
and salts this with the commit message. Added the CLI option
to override the default behaviour (where only pre-commit.ci and
dependabot changes could be merged) so that on first run, the tool
will provide a SHA value representing the user's bulk pull requests.
Providing that on the CLI with --override will permit standard user
pull requests to also be merged.

Signed-off-by: Matthew Watkins <mwatkins@linuxfoundation.org>