Feat: Improve progress display and feedback
Also, adds scan option to list all the unmergeable pull requests in
the GitHub orgnisation, with a view to directing the user to those
needing attention or revisions, such as rebase operations, or where
linting failures have been flagged. Also, added a table column in the output
showing the number of copilot comments that have not been resolved.

Signed-off-by: Matthew Watkins <mwatkins@linuxfoundation.org>