Fix: Rename dry-run -> preview, fix --force behaviour
The --force option parameters were not making their way down the calling stack
to the merge-manager correctly, meaning merge operations that should have
succeeded were still failing. This contains a fix which means merges now work
as expected. Also, since the removal of the previous --dry-run CLI flag/arg,
we have references to this all over the code base. Since we have changed the
behaviour of the tool, this needed to be renamed "preview" to reflect how the
tool actually works. The documentation and code base are now consistent with
the way the implementation works in practice.

Signed-off-by: Matthew Watkins <mwatkins@linuxfoundation.org>